Profiler: extract Profile/Profiler protocols, rename concrete types to Time*
The upcoming AllocationProfiler will share the begin/end/write public
surface with the existing time-sampling profiler. Hoist that shared
contract into protocols so call sites that want to drive profilers
generically can hold any Profiler/any Profile, with kind-specific data
(samples, allocation events, etc.) accessed by downcast.

Naming follows Apple Instruments' subject-based convention (Time
Profiler, Allocations): TimeProfiler/TimeProfile now, AllocationProfiler/
AllocationProfile later. ProfileMetrics becomes TimeProfileMetrics since
capture-timing percentiles are time-sampling-specific; an allocation
profile will have its own metrics shape.

TimeProfiler.main stays on the concrete class rather than moving onto
the protocol: allocation profiling is process-wide, not thread-bound,
so a single shared instance there would not make sense.

Source-breaking: no. The KSCrashProfiler module was added after the
2.5.0 release tag, so no symbol this rename touches existed at the
most recent release.
